Refactor input and output handling.
When no output file is specified, it is assumed to be stdout.
When no input file is specified, it is assumed to be stdin.

This change allows svgmin to be treated like normal Unix tool, i.e. piping
and redirecting will just work, useful for e.g. multiple filtering and/or
in combination with other tools.